src/com/zjxjwxk/leetcode/_0041_First_Missing_Positive/Solution.java
package com.zjxjwxk.leetcode._0041_First_Missing_Positive;

/**
 * 置换
 * @author Xinkang Wu
 * @date 2022/1/22 14:54
 */
public class Solution {

    public int firstMissingPositive(int[] nums) {
        int len = nums.length;
        for (int i = 0; i < len; ++i) {
            while (nums[i] > 0 && nums[i] <= len && nums[i] != nums[nums[i] - 1]) {
                swap(nums, i, nums[i] - 1);
            }
        }
        for (int i = 0; i < len; ++i) {
            if (nums[i] != i + 1) {
                return i + 1;
            }
        }
        return len + 1;
    }

    private void swap(int[] arr, int index1, int index2) {
        int temp = arr[index1];
        arr[index1] = arr[index2];
        arr[index2] = temp;
    }
}
